puppetdb-tooquiet - make the 'role' fact configurable
The default is to print the fact 'role' to print on the end of puppet,
but it's nice to make that configurable.

Thanks to Jon Robertson, https://github.com/jonrober/cms-puppetdb-tools/

Also, the formatting is a bit different (more characters for the
hostname, less for the role name; we probably want that to be
configurable too).